Update es5-shim.js
This check breaks ie8.

Test case: [undefined].indexOf(undefined) returns -1 rather than 0

The reason for this is because in ie8

0 in [undefined] returns false, whereas in more modern browsers it returns true.

I'm not sure why this check was being done in the first place and think it may not be required (hence my proposed change).  If it is required, can we find a way to achieve the same thing without breaking ie8?