Update String#trim spec compliance
Test for incorrect trimming of NEL, trim ZWSP (which *is* trimmable in the oldest Unicode spec ES5 targets), reduce the number of string concatenations, and more quickly determine `hasTrimWhitespaceBug` to be false in the usual pre-ES5 case.

I would change `if (typeof this === 'undefined' || this === null)` to `if (this == null)` but I don't know whether that's allowed by this library's style guide.