Fix Object.create sham to not set __proto__
Background:
Object.create(null) is great, because it creates objects whose key
space is not polluted by names like "toString" and "constructor".
In addition, in modern browsers, objects created with
Object.create(null) do not even have a `__proto__` property!  This means
`'__proto__' in x` is false and `x.__proto__ === undefined` when
`x = Object.create(null)`.  Browsers extend this courtesy even further;
if you even set `x.__proto__ = null` at any time, the `__proto__`
property of x loses its special meaning, and from then on you can assign
x's `__proto__` property without actually changing the internal
prototype pointer of x.

Note that older versions of modern engines, in particular Node 0.10,
are not quite so courteous and will return true for `'__proto__' in x`
and report a `__proto__` of null.  I'm not sure if the spec says
anything about this (the ES5 spec doesn't mention `__proto__`, but the
ES6 spec apparently does, so maybe it had something to do with the move
towards eliminating `__proto__` on "nullary" objects).  Either way,
having an enumerable `__proto__` property on x is definitely wrong.

The Object.create sham uses a very clever technique to create nullary
objects in browsers that don't support `__proto__` (like IE <11).
However, it then turns around and sets a property named `__proto__` on
the object!  Note that the code has already determined that this
property has no special meaning in the current engine, so it only gets
in the way.

This commit removes the assignment to `__proto__` in those browsers, so
that `'__proto__' in x` is false and `x.__proto__ === undefined`, and
no `__proto__` property is encountered when enumerating over x.

It also adds a check to `Object.getPrototypeOf` so that
`Object.getPrototypeOf(Object.create(null))` continues to return null.
This new check causes `Object.getPrototypeOf` on cross-realm objects
(such as objects from iframes) to return null instead of
`Object.prototype`, but neither is correct and we are just being best-
effort in this case (there is no way to get the real answer).

See #301.