Check for module's existence instead of exports
The library should be checking for `module`'s existence instead of `exports` existence. It has happened to us that in the browser for whatever reason `exports` was pointing to `Window` and the module check failed on `Module is not defined` on line 25 in this file. Also it doesn't make sense to check for variable `A` and then writing to variable `B`.